Return queue, external command input
Created the return queue, a queue that holds pointers to locations where the return values of commands will be placed once they're executed. Also reworked the external command input, making enqueue externally accessible for use. It now returns a void pointer to the location of the eventual return value. NOTE: This return value memory must be freed by the caller.